requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.13
django-ckeditor-5==0.2.13
django-heroku==0.3.1
django-js-asset==2.2.0
django-tinymce==4.0.0
fontawesomefree==6.5.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.0
whitenoise==6.6.0
django-tinymce==4.0.0

